qa/tasks/salt: change deepsea.state into salt.command
deepsea.state was not specific to deepsea - it seemed to be trying
to be a way to run any salt command. So let's make it do that instead.

This has the side effect that we can use it to run saltutil.sync_all, as
well.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit 51ec1e95e8cd43acb42c14e740e8a8961b802d06)

Conflicts:
	qa/tasks/deepsea.py